fix timer bug
The timer was starting as soon as the page loaded, so I set the interval to null initially so that the startTimer function would only run when the clickedCard event fires, and will not re-run when the event fires again. Also moved the startTimer function to the cardClicked event to prevent the user starting the game when the deck container has been clicked.